requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyJWT==1.7.1
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
twilio==6.53.0
urllib3==1.26.7
Werkzeug==2.0.2
